fix: move styled html elements into styledHTMLElements folder
