+++ src/components/styledHTMLElements/StyledListItem.tsx
@@ -1,0 +1,11 @@
+import { LiHTMLAttributes } from "react"
+
+type ListItemProps = LiHTMLAttributes<HTMLLIElement>
+
+export const ListItem = (props: ListItemProps) => {
+	return (
+		<li className="flex h-12 w-full cursor-pointer rounded-lg p-2  hover:bg-gray-50">
+			{props.children}
+		</li>
+	)
+}
